requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
Werkzeug==3.1.3
